Rename Manager files and class name
