+++ lib/serviceStateManager/ServiceStateManager.js
@@ -3,7 +3,7 @@ const util = require('util');
 
 const { Logger } = require('../logging/Logger');
 
-class Manager {
+class ServiceStateManager {
   /**
    * Manages the services' states, providing health check and shutdown utilities.
    *
@@ -16,7 +16,7 @@ class Manager {
      * @private
      */
     this.logger = new Logger('microservice-sdk:service-state-manager');
-    this.logger.info('Initializing the Manager...');
+    this.logger.info('Initializing the ServiceStateManager...');
 
     /**
      * @type {Map<string, {interval: NodeJS.Timeout, inUse: boolean, status: boolean}>}
@@ -246,4 +246,4 @@ class Manager {
   }
 }
 
-module.exports = Manager;
+module.exports = ServiceStateManager;
